Implement PortMapper component
This component is responsible for distributing L4+ services among
dynamically allocated (via a replacable backend) layer 3 ports
(IP addresses).

It is designed in such a way that it can be interrupted at any time
and it will converge to a sane state upon a full resync.

This is achieved by tracking the OpenStack ports assinged to
services as annotations (to be able to recover the
non-deterministic parts of the service->port assignment) as well
as recovering gracefully if conflicts arise on that.