Annotate Service objects with Ingress info
This was harder than expected, since the LoadBalancer status and
the annotations cannot be updated in the same API call (or even
in directly subsequent calls, since changing either changes the
resource revision and the resource would’ve to be fetched
in-between).

In order to make this safe against races and crashes, we decided
to use the k8s update notifications as a driver for incrementally
converging the state.

That has a few implications:

- the IP address information is not assigned immediately
- we need extra round-trips

However, it doesn’t seem to be prohibitively slow.