Prefer in-memory state over Service state
This is useful in situations where the Annotation can not be
updated, for example because the LoadBalancer Status field needs
to be cleared first.

In this case, the annotation may either not be set at all or
it may not be up to date with our internal state. If we do not
prefer the internal state over the annotation, we may end up
allocating a fresh port instead of reusing the port already
provisioned for the service.