controller: return error if fip creation failed
If for some reason the creation of a floating IP (fip) in OpenStack fails it
should be made clear this is unexpected behaviour. One obvious reason
for a failed fip creation are quotas. If no fip was provided, this led
to the following execution flow:

- fip creation failed
- port is deleted
- "" is returned as port ID
- in the next round, the controller wants to look up the port with
  non-existing port with the zero-length ID and crashes
- horrible things happen from this stage on

By returning an error none of the above will happen and the pending
service(s) are re-queued for port/fip creation in the next iteration.