Rip out the port cache
For a while now (potentially since ever), ch-k8s-lbaas has been known
for losing floating IPs. We were so far unable to track it down---the
logs were always inconclusive and it wasn't fully clear where the issue
might come from.

As "cache invalidation" is one of the two hard problems in computer
science (next to naming things and off-by-one errors), it seems most
sensible to rip out the port cache.

With the typical cluster size, it should not be necessary to have a
cache. In addition, using a cache is in direct contradiction to acting
on the current state (even though API calls also always return stale
data, but at least it's less stale than your average cache), which we
normally want to do in the Kubernetes world.

So let's try this, for a change.