Add NetworkPolicies and PolicyAssignments to model
This commit adds two new elements to the second hierarchy level of the model:
NetworkPolicies and PolicyAssignments.

Each NetworkPolicy entry can contain zero or more AllowedIngresses which in
turn can contain zero or more IPBlockFilters and zero or more PortFilters.
namespaceSelectors and podSelectors are not part of the model.

PolicyAssignments contains a mapping from one podIP to zero or more
NetworkPolicy names.

The new entries in the model do not interfere with the existing Ingress
element. This accounts for the fact that "Pod is matched by a Service" and
"Pod is matched by a NetworkPolicy" are orthogonal.