Implement NetworkPolicy handling in controller
This commit adds support to the controller for creating NetworkPolicy and
PolicyAssignment entries in the model.

For each NetworkPolicy in the cluster which applies to Ingress
(spec.PolicyTypes contains "Ingress"), a NetworkPolicy entry in the
cluster is created. All IngressRules that contain an ipBlock are added
to the NetworkPolicy entry as AllowedIngress. Policies that only contain
namespaceSelectors or podSelectors result in a NetworkPolicy entry
without AllowedIngress.

Also, for each NetworkPolicy in the cluster which applies to Ingress,
and for each pod to which it applies, an entry in PolicyAssignment is
created.