Add support for PolicyAssignments to agent
For each podIP for which a PolicyAssignment exists in the model,
* a new chain in the filter table with default 'drop' is created.
  For each NetworkPolicy assigned to this podIP, a jump rule is added
  that points to the chain of the respective NetworkPolicy.
* a 'goto' rule is added to the forward chain that points to the podIP
  specific chain.

The existing code to generate the nat table remains untouched.
NetworkPolicies are handled after the DNAT because only then is the
podIP known. This accounts for the fact that "Pod is matched by the Service"
and "Pod is matched by the NetworkPolicy" are orthogonal.