Add input validation to the agent
This commit adds validate annotations to the loadbalancer model and a
validation to the agent. In case the received JSON is not valid, the
agent returns HTTP status 400.

This is preferable over manually checking the struct fields because (1)
it provides a single source of truth (also for the controller and other
implementations of the agend) and (2) it ensures we didn't miss anything.
Individual functions may still perform (redundant) sanity checks.

Fixes #21