Add port IDs and additional IPs to agent config
These values are used by the OpenStackL3PortManager to create a VRRP
setup that is compatible with new versions of the OVN ML2 plugin and
need to be added to the configuration with yaook/k8s.
Additionally, the openstack config was moved to the config package to
avoid cyclic imports.