Create VRRP setup for OVN-based OpenStack
To setup VRRP with new versions of the OVN ML2 plugin, additional IP
addresses of the agent configuration (at the moment of this commit only
the IP address of the VIP port is used) and all fixed IP addresses of
all L3 ports need to be added to the allowed address pairs of each agent
node. This is done by overwriting the address pairs accordingly after
the provisioning or deletion of a L3 port as well as periodically during
the entire runtime of the k8s-lbaas using the EnsureAgentsStateJob. The
periodic job is used to ensure the setup will still be created after an
agent was unreachable.

As a part of this commit functions provided by the package
openstack/networking/v2/ports were wrapped into the PortClient to allow
better testability.